frontend/src/services/aiAgentService.ts
import { AgentType, AgentTask, AnalysisResult } from '../types/aiTypes';

class AIAgentService {
  private async runAgent(type: AgentType, input: any): Promise<any> {
    // TODO: Replace with actual API call to your AI backend
    return new Promise((resolve) => {
      setTimeout(() => {
        switch (type) {
          case 'STRATEGIC_ANALYSIS':
            resolve({
              goals: ['Improve user experience', 'Increase efficiency'],
              marketAnalysis: 'Competitive market with growth potential',
              recommendations: ['Focus on UI/UX', 'Implement automation']
            });
            break;
          case 'REQUIREMENTS_ANALYSIS':
            resolve({
              functional: ['User authentication', 'Data visualization'],
              nonFunctional: ['Response time < 2s', '99.9% uptime'],
              constraints: ['Budget limitations', 'Timeline constraints']
            });
            break;
          default:
            resolve({});
        }
      }, 1000);
    });
  }

  async generateProjectPlan(description: string): Promise<AnalysisResult> {
    const tasks: AgentTask[] = [
      {
        id: '1',
        type: 'STRATEGIC_ANALYSIS',
        status: 'pending'
      },
      {
        id: '2',
        type: 'REQUIREMENTS_ANALYSIS',
        status: 'pending'
      }
    ];

    const results = await Promise.all(
      tasks.map(async (task) => {
        task.status = 'started';
        const result = await this.runAgent(task.type, description);
        task.status = 'completed';
        task.result = result;
        return result;
      })
    );

    // Convert agent results to node format
    return this.convertToNodes(results);
  }

  private convertToNodes(results: any[]): AnalysisResult {
    // This is a placeholder implementation
    return {
      requirements: [
        {
          id: 'req1',
          type: 'requirement',
          title: 'User Authentication',
          description: 'Implement secure user authentication system',
          priority: 'high',
          category: 'security'
        }
      ],
      architecture: [
        {
          id: 'arch1',
          type: 'architecture',
          title: 'Authentication Service',
          description: 'OAuth2 based authentication service',
          component: 'backend',
          dependencies: []
        }
      ],
      timeline: [
        {
          id: 'time1',
          type: 'timeline',
          title: 'Authentication Implementation',
          description: 'Implement user authentication system',
          duration: 14,
          dependencies: []
        }
      ],
      risks: [
        {
          id: 'risk1',
          type: 'risk',
          title: 'Security Vulnerability',
          description: 'Potential security vulnerabilities in authentication',
          impact: 'high',
          probability: 'low',
          mitigation: 'Regular security audits and penetration testing'
        }
      ]
    };
  }
}

export const aiAgentService = new AIAgentService();
